[HTML5] Change default resolution for webcams to 320x240
- This was done to mitigate packet loss and flickering and better abide to our default configured bitrate (300kbps).
This goes alongside updates to bbb-webrtc-sfu and kurento-media-server to increase reliablity and diminish stream artifacts.
I'll probably revisit this resolution change once I can correctly identify the source of corrupted frames, but I really think we should
maintain this as 320x240 for the time being. 640x480 would only be feasible if we increased our default bitrate config to 500kbps.