Don't recommend Firefox or Chrome for WebRTC when Puffin is being used
People use Puffin Browser only when they are on a mobile device and have to use
BBB flash client. They cannot use a browser with WebRTC support,
in our case, we in Dumalogiya temporarily disabled HTML5 version
due to problems with sharing webcams across HTML5 and Flash clients and we are
not ready to fully migrate to HTML5 client because it e.g. lacks support
of changing window layouts. So, we want to avoid showing this warning in Puffin Browser.