Automate Conflict Check for Pull Requests
This patch adds a simple GitHub Actions workflow which will
automatically check open pull requests on code changes, adding a comment
and a label if the pull request has conflicts. The label will
automatically be removed again once the conflict has been resolved.

We use this for a while now with opencast and it really helps since
people do not have to manually check if their pull requests need an
update any longer.

An example, of hoe this looks in action can be found at
https://github.com/opencast/opencast/pull/2053#issuecomment-734987582