let bbb-apps-akka read an override config file from /etc/bigbluebutton
This change introduces a config file
`/etc/bigbluebutton/bbb-apps-akka.conf` which reads the default config
from packages and allows operators to keep their own config file changes
across package upgrades. This removes the workaround of copying the
config file to /tmp in preinst and then extracvting API URI and secret
and then in postinst inject it back to the config file in /usr/share

If the config file in /usr does not have the "changeme" secret, then API
URI and secret are extracted and injected to the generated file /etc, so
packages would upgrade in a clean way.

The postrm script removes the config file in /etc if the package is
purged but not if it is just uninstalled.

bbb-install and bbb-conf must reflect this change and replace the API
secret and URL in /etc, not /usr/share