fix(polling): stop current running poll
As far as I could understand, polls are tightly coupled with the meeting
main content area, or at least they were and we still have to deal with
this close relation between them. Not sure if it's something we'll keep
this way forever but, from my candid perspective, looks like this is already
diverging inside the poll model. Polls are indexed by presentation pages,
screenshare or even something called "public" that I'm not 100% what
actually means. My best guess is anything besides the first and the second.

The polling stop process lacks to inform which pollId is scoped at source
so it relies on akka-apps to discover based on the current state of other
apps. This looks like the major problem over this polling termination issue.

Made a few changes at the running poll getter fallback at the polling stop
process. Following the premise that there is only one running poll available,
we make sure to return a valid running poll (if there's one).