Set worker_rlimit_nofile only once in nginx.conf
worker_rlimit_nofile may only be set once in nginx.conf. So far there was no check whether it is already set to a value different than 10000. This results in the option beeing set twice in the config file, if it was already set by. This caused nginx to fail startup. Now it is checked that the limit is at least 10000 and only if not it is set to 10000.