build: add /wasm location block to bbb-html5.nginx
Guarantees that nginx will be the one serving the WASM binaries (static assets).

The types block is there to guarantee that the correct MIME type is advertised.

Moreoever, it's preferable that the WASM binaries are gzipped because they're
too large to ignore, so it's worth the server-side gzipping overhead.
The explicit gzip type declaration is there because only text/html assets are
compressed by default.