feat: add new dialplan rule for bbb-webrtc-sfu calls
This new dialplan rule filters calls originating from bbb-webrtc-sfu via SIP
user agent parsing. The default bbb-webrtc-sfu UA is "bbb-webrtc-sfu".

A new dialplan rule is needed to force RTP auto-adjustment for calls originating
in bbb-webrtc-sfu (rtp_manual_bugs=ACCEPT_ANY_PACKETS).
That is due to the fact that bidirectional mediasoup bridging is done via an
RTP/AVPF endpoint which does not use ICE. FreeSWITCH arbitrarily blocks off auto
adjustment for AVPF profiles (presuming ICE), so it needs to be forced otherwise
the bridge won't work properly in all environments.

Bridging mediasoup and FS via WebRTC (which would circumvent that) is currently
not an option due to the fact that FreeSWITCH doesn't handle STUN role conflicts
properly (and there will always be a conflict since the initiator is controlled
and FS always defaults to controlled)

Briding mediasoup and FS via plain RTP/AVP (which would also circumvent that) is
not an option right now due to the fact that FreeSWITCH doesn't make ssrcs
public in signaling for RTP/AVP profiles. mediasoup needs the remote ssrcs.
This could work by pre-generating a ssrc in bbb-webrtc-sfu, signaling it via a
SIP header and then specifying it in the rtp_use_ssrc channel variable in FS,
which would allow us to shim the ssrc in FS's answer in bbb-webrtc-sfu.
Maybe in the future.