fix(tldraw): fix shapes not persisting in production +
Tldraw uses the constructor name for session complete names:
https://github.com/tldraw/tldraw/blob/80ed0bf27bbaa8566b230df318364dec85304d31/packages/tldraw/src/state/TldrawApp.ts#L2916
So in production they are changed because of js minifier.

I changed the onCommand reactor to be one generic for all session:complete actions.
Also fixed some crashes with grouping and ungrouping shapes, in those cases the order of adding/removing is important because children have a reference to the father (grouping shape).