Revert "fix: don't clear text fake annotations"
This reverts commit 6200b087f1fe01bbd184ca72af5c7cbe723641f3.
Reverting this commit because it causes fake annotations from the text
tool not to be removed. Since the text tool is the only one that
generates fake annotations(because it is the only one that is live
synced), the function to clear fake annotations is directly intended to
clear them.