Add `npm run test:ssh` command to test suite
This npm script will ssh to a BigBlueButton server, obtain its URL and SECRET,
then run the test suite using these settings.

This reduces the complexity of running tests (no need to set environment variables) if you have ssh access to the server.

The shell script is a bit obscure, and comments are not allowed in JSON.

The "set -e" switch is used to exit the script if the ssh fails, but this only works
if the ssh is not part of a pipeline, which is why it is assigned to a separate variable.
(see https://unix.stackexchange.com/a/23099/37949)