bbb-presentation-video: Update to 4.0.0-rc.1
See the release notes:
https://github.com/bigbluebutton/bbb-presentation-video/releases/tag/4.0.0-rc.1

This version should be feature-complete for BigBlueButton 2.6 whiteboard
video generation.

I have updated the `bbb-presentation-video.placeholder.sh` script to
download the prebuilt packages from the Github release rather than clone
the source code. Work to automate handling of including this package in
the builds automatically is still yet to be done.